back to js webpack configs bc esm incompatibility
