+++ webpack.dev.config.js
@@ -1,36 +1,33 @@
 import path from 'path';
-import webpack from 'webpack';
 import HtmlWebPackPlugin from 'html-webpack-plugin';
 import { fileURLToPath } from 'node:url'
+import ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin';
+import ESLintPlugin from 'eslint-webpack-plugin';
 
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
-const config: webpack.Configuration = {
-  devtool: 'source-map',
-  mode: 'production',
+const config = {
+  devtool: 'inline-source-map',
+  mode: 'development',
   entry: path.resolve(dirname, 'demo/index.tsx'),
   output: {
     filename: 'demo.bundle.js',
-    path: path.resolve(dirname, 'dist-demo'),
+    path: path.resolve(dirname, 'demo'),
   },
   module: {
     rules: [
       {
-        test: /\.(js|jsx|ts|tsx)$/,
-        exclude: /node_modules/,
-        use: [{
-          loader: 'ts-loader',
-          options: {
-            configFile: 'tsconfig.json',
-            compilerOptions: {
-              outDir: "./dist-demo",
-              declarationDir: undefined,
-              declaration: false
-            },
-          },
-        }],
-      },
+        test: /\.tsx?$/,
+        loader: 'ts-loader',
+        options: {
+          transpileOnly: true,
+          compilerOptions: {
+            module: "ESNext",
+            moduleResolution: "node",
+          }
+        },
+      }
     ],
   },
   resolve: {
@@ -44,10 +41,19 @@ const config: webpack.Configuration = {
     },
   },
   plugins: [
+    new ReactRefreshWebpackPlugin(),
     new HtmlWebPackPlugin({
       template: 'demo/index.html',
     }),
+    new ESLintPlugin({
+      fix: true,
+      emitWarning: true,
+    }),
   ],
+  devServer: {
+    port: 3000,
+    host: '0.0.0.0',
+  },
 };
 
 export default config;
